# Brain_API_Specification.yaml
openapi: 3.1.0
info:
  title: Brain API
  version: 0.3.0
  description: |
    The Brain API exposes six layers as HTTP endpoints plus one MCP surface.
    Raw (evidence), Ledger (normalized financial truth), Wiki (memory),
    Policy (governance), Agent (action orchestration), Audit (proof).

    v0.2.0 introduces the Normalized Ledger layer. The Wiki API is preserved
    but its role narrows to memory rendering and Q&A — financial truth lives
    in the Ledger. The Execution endpoints are renamed to Agent endpoints;
    legacy /execution/* paths remain shipped with deprecation headers for
    the duration of the v0.3 transition.

    v0.3.0 aligns the spec to the canonical public docs at
    https://docs.brain.fi. Material changes land in small commits — see
    docs/sdk-audit.md for the reconciliation queue. This release aligns
    the error envelope to the docs shape (`{ error: { code, message,
    details, trace_id, docs_url } }`); the v0.2 flat shape is replaced.

    All endpoints are tenant-scoped. Tenant identity is derived from the
    authenticated principal. The principal's tenant is never passed in URLs
    except where explicitly required for cross-tenant operations by platform
    administrators.

    Base URL (production): https://api.brain.fi/v1
    Base URL (sandbox):    https://api.brain.dev/v1
    Protocol: HTTPS only. HTTP requests are rejected at the edge.
    Content type: application/json unless otherwise specified.
    Character encoding: UTF-8.
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /sources:
    post:
      tags: [Sources]
      summary: Connect a new source
      operationId: connectSource
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenantId, type, credentials]
              properties:
                tenantId: { type: string }
                type: { $ref: '#/components/schemas/SourceType' }
                credentials:
                  type: object
                  additionalProperties: true
                  description: |
                    Provider-specific credentials. Schema depends on
                    `type` — e.g. Plaid expects `{access_token}`,
                    Stripe `{api_key}`. Server validates per adapter.
                metadata:
                  type: object
                  additionalProperties: true
                  description: |
                    Optional caller labels (e.g. `{label: "operating
                    account"}`). Stored verbatim and returned with the
                    Source record.
      responses:
        '201':
          description: Source connected
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Source' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401':
          description: Provider rejected credentials (`source_credential_invalid`)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

    get:
      tags: [Sources]
      summary: List sources
      operationId: listSources
      parameters:
        - name: tenantId
          in: query
          schema: { type: string }
        - name: type
          in: query
          schema: { $ref: '#/components/schemas/SourceType' }
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/SourceStatus' }
        - name: limit
          in: query
          schema: { type: integer, default: 50, maximum: 500 }
        - name: cursor
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Source list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Source' }
                  next_cursor:
                    type: string
                    nullable: true
  /sources/{source_id}:
    get:
      tags: [Sources]
      summary: Get a single source
      operationId: getSource
      parameters:
        - name: source_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Source
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Source' }
        '404': { $ref: '#/components/responses/NotFound' }

    delete:
      tags: [Sources]
      summary: Disconnect a source
      description: |
        Marks the source `disconnected`. Ingestion stops immediately;
        historical Raw artifacts are retained per the tenant's
        retention policy and remain queryable via /raw/*.
      operationId: disconnectSource
      parameters:
        - name: source_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Disconnected
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Source' }
        '404': { $ref: '#/components/responses/NotFound' }
  /sources/{source_id}/sync:
    post:
      tags: [Sources]
      summary: Trigger an immediate sync
      description: |
        Queues a manual sync. Returns the sync job descriptor; the job
        runs asynchronously and emits new `/raw/ingest` rows as data
        arrives. Stub-mode source types accept the call but return
        `{notes: "stub"}` instead of running real adapter logic.
      operationId: syncSource
      parameters:
        - name: source_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '202':
          description: Sync enqueued
          content:
            application/json:
              schema:
                type: object
                required: [job_id, status]
                properties:
                  job_id: { type: string }
                  source_id: { type: string }
                  status:
                    type: string
                    enum: [enqueued, running]
                  notes:
                    type: string
                    description: |
                      Present and set to `"stub"` when the underlying
                      adapter is not yet implemented (csv_upload,
                      pdf_upload, email_inbound, netsuite,
                      alchemy_wallet, eth_address in v0.3 ship).

  #############################################################################
  # LAYER 1: RAW
  #############################################################################
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT bearer token issued by Brain's auth service.
        Contains tenant_id and scopes. See Engineering Standards
        section 2 for the full auth model.
...
  responses:
    BadRequest:
      description: Request is malformed or missing required fields
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    NotFound:
      description: Resource not found or not accessible to this tenant
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [error]
      description: |
        Canonical Brain error envelope. Every non-2xx response wraps its
        body in this shape. Source of truth:
        https://docs.brain.fi/resources/errors (canonical wire codes) and
        https://docs.brain.fi/api-reference/overview (envelope shape).

        Engineering Standards §4.1 previously specified the same nested
        shape under the field name `request_id`; the docs publish it as
        `trace_id` and the docs win — see docs/sdk-audit.md conflict E.
      properties:
        error:
          type: object
          required: [code, message, trace_id, docs_url]
          properties:
            code:
              type: string
              description: |
                Machine-readable error code. Stable: once shipped a code
                is never renamed (Standards §4.3 "additions only"). See
                services/api/src/shared/errors.ts for the registry.
            message:
              type: string
              description: Human-readable error message (English, untranslated).
            details:
              type: object
              additionalProperties: true
              description: |
                Optional structured payload. Schema is code-specific —
                e.g. `policy_denied` carries `{rule_id, …}`, `gate_*`
                carries `{check, …}`.
            trace_id:
              type: string
              description: |
                Correlation id for support. Per-request, surfaced in
                logs, metrics, and traces. Pass to support to retrieve
                the full request timeline.
              example: trc_8f3a92...
            docs_url:
              type: string
              format: uri
              description: |
                Direct link to the docs page for this error code, of the
                form https://docs.brain.fi/errors/{code}.
              example: https://docs.brain.fi/errors/policy_denied
      example:
        error:
          code: policy_denied
          message: Counterparty not approved
          details:
            counterparty_id: cp_x
            policy_version: 3
          trace_id: trc_8f3a92abc
          docs_url: https://docs.brain.fi/errors/policy_denied

    #
    # Sources (PLAN-FIRST #12) — adapter connections that feed the Raw
    # layer. The 8-value MVP set per docs/sdk-audit.md decision K2.
    # Adding a new value to `SourceType` is a non-breaking enum widening
    # — extends as adapters ship (Architecture §3.1, "Other
    # source-specific adapters … are post-MVP").
    #
    SourceType:
      type: string
      description: |
        Source connector identifier. v0.3 MVP set (8 values).
        Stub-mode connectors: csv_upload, pdf_upload, email_inbound,
        netsuite, alchemy_wallet, eth_address. Concrete connectors:
        plaid, stripe.
      enum:
        - plaid
        - stripe
        - netsuite
        - email_inbound
        - csv_upload
        - pdf_upload
        - alchemy_wallet
        - eth_address
    SourceStatus:
      type: string
      enum:
        - active
        - paused
        - error
        - disconnected
    Source:
      type: object
      required: [id, tenantId, type, status, created_at]
      properties:
        id:
          type: string
          description: Source id (src_<ulid>).
        tenantId:
          type: string
        type:
          $ref: '#/components/schemas/SourceType'
        status:
          $ref: '#/components/schemas/SourceStatus'
        last_synced_at:
          type: string
          format: date-time
          nullable: true
        metadata:
          type: object
          additionalProperties: true
          description: |
            Caller-supplied labels (`label`, `description`, etc.).
            Stored verbatim and returned unchanged.
        error_message:
          type: string
          nullable: true
          description: |
            Most recent provider-level error. Populated when status =
            `"error"` (typically after `source_credential_invalid` or
            `upstream_timeout`).
        is_stub:
          type: boolean
          description: |
            True when the underlying adapter is not yet implemented and
            sync/ingest return a `{notes: "stub"}` shape. Useful for
            UIs that want to grey out "Sync now" for stub sources.
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...